requirements.txt
# Web scraping and HTTP requests
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Data processing
pandas==2.1.4
pydantic==2.5.2

# CLI and configuration
click==8.1.7
pyyaml==6.0.1

# Logging and utilities
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code quality and type checking
mypy==1.7.1
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Type stubs
pandas-stubs==2.1.4.231227
types-PyYAML==6.0.12.12
types-beautifulsoup4==4.12.0.7

# Optional: For future LLM integration
# openai==1.3.8
# anthropic==0.7.8
